+++ tests/regression/test_fors_with_var_iteration_counts.cpp
@@ -63,7 +63,7 @@ main(void)
     int a = 3;
 
     for (int i = 0; i < BUFFER_SIZE; i++) {
-        A[i] = i;
+        A[i] = (float)i;
     }
 
     for (int i = 0; i < WORK_ITEMS; i++) {
